Second Commit
-Added repository method tests
-Added current project to homepage
-Added a link on the homepage that links to a list of all sponsors and subsequently their info

+++ src/main/java/com/example/demo/repository/ItemRepository.java
@@ -8,5 +8,5 @@ import org.springframework.stereotype.Repository;
  * Created by John on 7/11/2017.
  */
 @Repository
-public interface ItemRepository extends CrudRepository<Item, Long>{
+public interface ItemRepository extends CrudRepository<Item, Integer>{
 }
